Hsy (#67)
* createReview complete

* review_rating

+++ src/main/java/com/TeamProject/TeamProject/Review/Review.java
@@ -20,7 +20,7 @@ public class Review {
   @Column(columnDefinition = "TEXT")
   private String content;
 
-  private Integer rating; //평점
+  private double rating; //평점
 
   private Integer thumbsup; // 해당 리뷰가 받은 추천수
